replace file for class structure
